Migrate rich-text image upload to chunked media endpoint
